SimpleClient -> KoraClient
