+++ experimental/s3-client-annotation-processor/src/test/java/ru/tinkoff/kora/s3/client/annotation/processor/S3KoraReactorClientTests.java
@@ -4,21 +4,20 @@ import org.junit.jupiter.api.Test;
 import ru.tinkoff.kora.annotation.processor.common.AbstractAnnotationProcessorTest;
 
 import java.util.List;
-import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
+class S3KoraReactorClientTests extends AbstractAnnotationProcessorTest {
 
     @Override
     protected String commonImports() {
         return super.commonImports() + """
+            import reactor.core.publisher.Mono;
             import java.nio.ByteBuffer;
             import java.io.InputStream;
             import java.util.List;
             import java.util.Collection;
-            import java.util.Optional;
             import ru.tinkoff.kora.s3.client.annotation.*;
             import ru.tinkoff.kora.s3.client.annotation.S3.*;
             import ru.tinkoff.kora.s3.client.model.*;
@@ -26,6 +25,24 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             import ru.tinkoff.kora.s3.client.model.S3Object;
             import software.amazon.awssdk.services.s3.model.*;
             """;
+    }
+
+    @Test
+    public void clientConfig() {
+        this.compile(List.of(new S3ClientAnnotationProcessor()), """
+            @S3.Client("my")
+            public interface Client {
+                        
+                @S3.Get
+                Mono<S3ObjectMeta> get(String key);
+            }
+            """);
+        this.compileResult.assertSuccess();
+        var clazz = this.compileResult.loadClass("$Client_Impl");
+        assertThat(clazz).isNotNull();
+
+        var config = this.compileResult.loadClass("$Client_ClientConfigModule");
+        assertThat(config).isNotNull();
     }
 
     // Get
@@ -36,22 +53,7 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.Get
-                S3ObjectMeta get(String key);
-            }
-            """);
-        this.compileResult.assertSuccess();
-        var clazz = this.compileResult.loadClass("$Client_Impl");
-        assertThat(clazz).isNotNull();
-    }
-
-    @Test
-    public void clientGetMetaOptional() {
-        this.compile(List.of(new S3ClientAnnotationProcessor()), """
-            @S3.Client("my")
-            public interface Client {
-                        
-                @S3.Get
-                Optional<S3ObjectMeta> get(String key);
+                Mono<S3ObjectMeta> get(String key);
             }
             """);
         this.compileResult.assertSuccess();
@@ -66,22 +68,7 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.Get
-                S3Object get(String key);
-            }
-            """);
-        this.compileResult.assertSuccess();
-        var clazz = this.compileResult.loadClass("$Client_Impl");
-        assertThat(clazz).isNotNull();
-    }
-
-    @Test
-    public void clientGetObjectOptional() {
-        this.compile(List.of(new S3ClientAnnotationProcessor()), """
-            @S3.Client("my")
-            public interface Client {
-                        
-                @S3.Get
-                Optional<S3Object> get(String key);
+                Mono<S3Object> get(String key);
             }
             """);
         this.compileResult.assertSuccess();
@@ -96,7 +83,7 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.Get
-                List<S3ObjectMeta> get(Collection<String> keys);
+                Mono<List<S3ObjectMeta>> get(Collection<String> keys);
             }
             """);
         this.compileResult.assertSuccess();
@@ -111,7 +98,7 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.Get
-                List<S3Object> get(List<String> keys);
+                Mono<List<S3Object>> get(List<String> keys);
             }
             """);
         this.compileResult.assertSuccess();
@@ -126,7 +113,7 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.Get("{key1}-{key2}")
-                S3ObjectMeta get(String key1, long key2);
+                Mono<S3ObjectMeta> get(String key1, long key2);
             }
             """);
         this.compileResult.assertSuccess();
@@ -141,7 +128,7 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.Get("{key1}-{key12345}")
-                S3ObjectMeta get(String key1);
+                Mono<S3ObjectMeta> get(String key1);
             }
             """));
     }
@@ -153,52 +140,7 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.Get("const-key")
-                S3ObjectMeta get();
-            }
-            """);
-        this.compileResult.assertSuccess();
-        var clazz = this.compileResult.loadClass("$Client_Impl");
-        assertThat(clazz).isNotNull();
-    }
-
-    @Test
-    public void clientGetKeyPrefix() {
-        this.compile(List.of(new S3ClientAnnotationProcessor()), """
-            @S3.Client("my")
-            public interface Client {
-                        
-                @S3.Get("pre-{key1}")
-                S3ObjectMeta get(String key1);
-            }
-            """);
-        this.compileResult.assertSuccess();
-        var clazz = this.compileResult.loadClass("$Client_Impl");
-        assertThat(clazz).isNotNull();
-    }
-
-    @Test
-    public void clientGetKeySuffix() {
-        this.compile(List.of(new S3ClientAnnotationProcessor()), """
-            @S3.Client("my")
-            public interface Client {
-                        
-                @S3.Get("{key1}-suffix")
-                S3ObjectMeta get(String key1);
-            }
-            """);
-        this.compileResult.assertSuccess();
-        var clazz = this.compileResult.loadClass("$Client_Impl");
-        assertThat(clazz).isNotNull();
-    }
-
-    @Test
-    public void clientGetKeyPrefixSuffix() {
-        this.compile(List.of(new S3ClientAnnotationProcessor()), """
-            @S3.Client("my")
-            public interface Client {
-                        
-                @S3.Get("pre-{key1}-suffix")
-                S3ObjectMeta get(String key1);
+                Mono<S3ObjectMeta> get();
             }
             """);
         this.compileResult.assertSuccess();
@@ -213,7 +155,7 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.Get("const-key")
-                S3ObjectMeta get(String key);
+                Mono<S3ObjectMeta> get(String key);
             }
             """));
     }
@@ -226,22 +168,7 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.List
-                S3ObjectMetaList list();
-            }
-            """);
-        this.compileResult.assertSuccess();
-        var clazz = this.compileResult.loadClass("$Client_Impl");
-        assertThat(clazz).isNotNull();
-    }
-
-    @Test
-    public void clientListObjects() {
-        this.compile(List.of(new S3ClientAnnotationProcessor()), """
-            @S3.Client("my")
-            public interface Client {
-                        
-                @S3.List
-                S3ObjectList list();
+                Mono<S3ObjectMetaList> list();
             }
             """);
         this.compileResult.assertSuccess();
@@ -256,7 +183,22 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.List
-                S3ObjectMetaList list(String prefix);
+                Mono<S3ObjectMetaList> list(String prefix);
+            }
+            """);
+        this.compileResult.assertSuccess();
+        var clazz = this.compileResult.loadClass("$Client_Impl");
+        assertThat(clazz).isNotNull();
+    }
+
+    @Test
+    public void clientListMetaFutureWithPrefix() {
+        this.compile(List.of(new S3ClientAnnotationProcessor()), """
+            @S3.Client("my")
+            public interface Client {
+                        
+                @S3.List
+                Mono<S3ObjectMetaList> list(String prefix);
             }
             """);
         this.compileResult.assertSuccess();
@@ -271,7 +213,7 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.List
-                S3ObjectList list(String prefix);
+                Mono<S3ObjectList> list(String prefix);
             }
             """);
         this.compileResult.assertSuccess();
@@ -280,13 +222,13 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
     }
 
     @Test
-    public void clientListLimitWithPrefix() {
+    public void clientListLimit() {
         this.compile(List.of(new S3ClientAnnotationProcessor()), """
             @S3.Client("my")
             public interface Client {
                         
                 @S3.List(limit = 100)
-                S3ObjectList list(String prefix);
+                Mono<S3ObjectList> list(String prefix);
             }
             """);
         this.compileResult.assertSuccess();
@@ -301,7 +243,7 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.List("{key1}-{key2}")
-                S3ObjectList list(String key1, long key2);
+                Mono<S3ObjectList> list(String key1, long key2);
             }
             """);
         this.compileResult.assertSuccess();
@@ -316,7 +258,7 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.List(value = "{key1}", delimiter = "/")
-                S3ObjectList list(String key1);
+                Mono<S3ObjectList> list(String key1);
             }
             """);
         this.compileResult.assertSuccess();
@@ -331,7 +273,7 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.List("{key1}-{key12345}")
-                S3ObjectList list(String key1);
+                Mono<S3ObjectList> list(String key1);
             }
             """));
     }
@@ -343,7 +285,7 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.List("const-key")
-                S3ObjectList list();
+                Mono<S3ObjectList> list();
             }
             """);
         this.compileResult.assertSuccess();
@@ -358,7 +300,7 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.List("const-key")
-                S3ObjectList list(String key);
+                Mono<S3ObjectList> list(String key);
             }
             """));
     }
@@ -371,7 +313,7 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.Delete
-                void delete(String key);
+                Mono<Void> delete(String key);
             }
             """);
         this.compileResult.assertSuccess();
@@ -386,7 +328,7 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.Delete("{key1}-{key2}")
-                void delete(String key1, long key2);
+                Mono<Void> delete(String key1, long key2);
             }
             """);
         this.compileResult.assertSuccess();
@@ -401,7 +343,7 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.Delete("{key1}-{key12345}")
-                void delete(String key1);
+                Mono<Void> delete(String key1);
             }
             """));
     }
@@ -413,7 +355,7 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.Delete("const-key")
-                void delete();
+                Mono<Void> delete();
             }
             """);
         this.compileResult.assertSuccess();
@@ -428,7 +370,7 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.Delete("const-key")
-                void delete(String key);
+                Mono<Void> delete(String key);
             }
             """));
     }
@@ -441,7 +383,7 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.Delete
-                void delete(List<String> key);
+                Mono<Void> delete(List<String> key);
             }
             """);
         this.compileResult.assertSuccess();
@@ -457,7 +399,7 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.Put
-                void put(String key, S3Body body);
+                Mono<Void> put(String key, S3Body body);
             }
             """);
         this.compileResult.assertSuccess();
@@ -472,7 +414,7 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.Put
-                S3ObjectUpload put(String key, S3Body body);
+                Mono<S3ObjectUpload> put(String key, S3Body body);
             }
             """);
         this.compileResult.assertSuccess();
@@ -487,7 +429,7 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.Put
-                void put(String key, byte[] body);
+                Mono<Void> put(String key, byte[] body);
             }
             """);
         this.compileResult.assertSuccess();
@@ -502,7 +444,7 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.Put
-                void put(String key, ByteBuffer body);
+                Mono<Void> put(String key, ByteBuffer body);
             }
             """);
         this.compileResult.assertSuccess();
@@ -517,7 +459,7 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.Put(type = "type")
-                void put(String key, S3Body body);
+                Mono<Void> put(String key, S3Body body);
             }
             """);
         this.compileResult.assertSuccess();
@@ -532,7 +474,7 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.Put(encoding = "encoding")
-                void put(String key, S3Body body);
+                Mono<Void> put(String key, S3Body body);
             }
             """);
         this.compileResult.assertSuccess();
@@ -547,7 +489,7 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.Put(type = "type", encoding = "encoding")
-                void put(String key, S3Body body);
+                Mono<Void> put(String key, S3Body body);
             }
             """);
         this.compileResult.assertSuccess();
@@ -562,7 +504,7 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.Put("{key1}-{key2}")
-                void put(String key1, long key2, S3Body body);
+                Mono<Void> put(String key1, long key2, S3Body body);
             }
             """);
         this.compileResult.assertSuccess();
@@ -577,7 +519,7 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.Put("{key1}-{key12345}")
-                void put(String key1, S3Body body);
+                Mono<Void> put(String key1, S3Body body);
             }
             """));
     }
@@ -589,7 +531,7 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.Put("const-key")
-                void put(S3Body body);
+                Mono<Void> put(S3Body body);
             }
             """);
         this.compileResult.assertSuccess();
@@ -604,7 +546,7 @@ class S3SimpleClientTests extends AbstractAnnotationProcessorTest {
             public interface Client {
                         
                 @S3.Put("const-key")
-                void put(String key, S3Body body);
+                Mono<Void> put(String key, S3Body body);
             }
             """));
     }
