+++ experimental/s3-client-symbol-processor/src/main/kotlin/ru/tinkoff/kora/s3/client/symbol/processor/S3ClientSymbolProcessor.kt
@@ -45,8 +45,8 @@ class S3ClientSymbolProcessor(
         private val CLASS_CONFIG = ClassName("ru.tinkoff.kora.s3.client", "S3Config")
         private val CLASS_AWS_CONFIG = ClassName("ru.tinkoff.kora.s3.client.aws", "AwsS3ClientConfig")
         private val CLASS_CLIENT_CONFIG: ClassName = ClassName("ru.tinkoff.kora.s3.client", "S3ClientConfig")
-        private val CLASS_CLIENT_SIMPLE_SYNC: ClassName = ClassName("ru.tinkoff.kora.s3.client", "S3SimpleClient")
-        private val CLASS_CLIENT_SIMPLE_ASYNC: ClassName = ClassName("ru.tinkoff.kora.s3.client", "S3SimpleAsyncClient")
+        private val CLASS_CLIENT_SIMPLE_SYNC: ClassName = ClassName("ru.tinkoff.kora.s3.client", "S3KoraClient")
+        private val CLASS_CLIENT_SIMPLE_ASYNC: ClassName = ClassName("ru.tinkoff.kora.s3.client", "S3KoraAsyncClient")
         private val CLASS_CLIENT_AWS_SYNC: ClassName = ClassName("software.amazon.awssdk.services.s3", "S3Client")
         private val CLASS_CLIENT_AWS_ASYNC: ClassName = ClassName("software.amazon.awssdk.services.s3", "S3AsyncClient")
         private val CLASS_CLIENT_AWS_ASYNC_MULTIPART = ClassName("software.amazon.awssdk.services.s3.internal.multipart", "MultipartS3AsyncClient")
